chore: benchmark revision-heavy eviction workloads
